+++ Src/PerformanceCollector/Shared/Implementation/QuickPulse/QuickPulseConstants.cs
@@ -1,0 +1,18 @@
+﻿namespace Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.Implementation.QuickPulse
+{
+    /// <summary>
+    /// Constants related to quick pulse service.
+    /// </summary>
+    internal class QuickPulseConstants
+    {
+        /// <summary>
+        /// Subscribed header.
+        /// </summary>
+        internal const string XMsQpsSubscribedHeaderName = "x-ms-qps-subscribed";
+
+        /// <summary>
+        /// Transmission time header.
+        /// </summary>
+        internal const string XMsQpsTransmissionTimeHeaderName = "x-ms-qps-transmission-time";
+    }
+}