+++ Src/PerformanceCollector/Unit.Tests/QuickPulse/QuickPulseServiceClientTests.cs
@@ -56,12 +56,12 @@
 
             this.pingResponse = response =>
                 {
-                    response.AddHeader("x-ms-qps-subscribed", true.ToString());
+                    response.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, true.ToString());
                 };
 
             this.submitResponse = response =>
                 {
-                    response.AddHeader("x-ms-qps-subscribed", true.ToString());
+                    response.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, true.ToString());
                 };
 
             string uriPrefix = string.Format(CultureInfo.InvariantCulture, "http://localhost:{0}/", Port);
@@ -280,7 +280,7 @@
             var serviceClient = new QuickPulseServiceClient(this.serviceEndpoint, string.Empty, string.Empty, string.Empty, string.Empty, new Clock(), false);
 
             // ACT
-            this.pingResponse = r => { r.AddHeader("x-ms-qps-subscribed", true.ToString()); };
+            this.pingResponse = r => { r.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, true.ToString()); };
             bool? response = serviceClient.Ping(string.Empty, DateTimeOffset.UtcNow);
 
             // ASSERT
@@ -296,7 +296,7 @@
             var serviceClient = new QuickPulseServiceClient(this.serviceEndpoint, string.Empty, string.Empty, string.Empty, string.Empty, new Clock(), false);
 
             // ACT
-            this.pingResponse = r => { r.AddHeader("x-ms-qps-subscribed", false.ToString()); };
+            this.pingResponse = r => { r.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, false.ToString()); };
             bool? response = serviceClient.Ping(string.Empty, DateTimeOffset.UtcNow);
 
             // ASSERT
@@ -312,7 +312,7 @@
             var serviceClient = new QuickPulseServiceClient(this.serviceEndpoint, string.Empty, string.Empty, string.Empty, string.Empty, new Clock(), false);
 
             // ACT
-            this.pingResponse = r => { r.AddHeader("x-ms-qps-subscribed", "bla"); };
+            this.pingResponse = r => { r.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, "bla"); };
             bool? response = serviceClient.Ping(string.Empty, DateTimeOffset.UtcNow);
 
             // ASSERT
@@ -344,7 +344,7 @@
             var serviceClient = new QuickPulseServiceClient(this.serviceEndpoint, string.Empty, string.Empty, string.Empty, string.Empty, new Clock(), false);
 
             // ACT
-            this.submitResponse = r => { r.AddHeader("x-ms-qps-subscribed", true.ToString()); };
+            this.submitResponse = r => { r.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, true.ToString()); };
             bool? response = serviceClient.SubmitSamples(new QuickPulseDataSample[] { }, string.Empty);
 
             // ASSERT
@@ -360,7 +360,7 @@
             var serviceClient = new QuickPulseServiceClient(this.serviceEndpoint, string.Empty, string.Empty, string.Empty, string.Empty, new Clock(), false);
 
             // ACT
-            this.submitResponse = r => { r.AddHeader("x-ms-qps-subscribed", false.ToString()); };
+            this.submitResponse = r => { r.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, false.ToString()); };
             bool? response = serviceClient.SubmitSamples(new QuickPulseDataSample[] { }, string.Empty);
 
             // ASSERT
@@ -376,7 +376,7 @@
             var serviceClient = new QuickPulseServiceClient(this.serviceEndpoint, string.Empty, string.Empty, string.Empty, string.Empty, new Clock(), false);
 
             // ACT
-            this.submitResponse = r => { r.AddHeader("x-ms-qps-subscribed", "bla"); };
+            this.submitResponse = r => { r.AddHeader(QuickPulseConstants.XMsQpsSubscribedHeaderName, "bla"); };
             bool? response = serviceClient.SubmitSamples(new QuickPulseDataSample[] { }, string.Empty);
 
             // ASSERT
@@ -671,7 +671,7 @@
                     this.pingResponse(context.Response);
 
                     var dataPoint = (MonitoringDataPoint)serializerDataPoint.ReadObject(context.Request.InputStream);
-                    var transmissionTime = long.Parse(context.Request.Headers["x-ms-qps-transmission-time"], CultureInfo.InvariantCulture);
+                    var transmissionTime = long.Parse(context.Request.Headers[QuickPulseConstants.XMsQpsTransmissionTimeHeaderName], CultureInfo.InvariantCulture);
 
                     this.pings.Add(Tuple.Create(new DateTimeOffset(transmissionTime, TimeSpan.Zero), dataPoint));
 
@@ -686,7 +686,7 @@
                     this.submitResponse(context.Response);
 
                     var dataPoints = serializerDataPointArray.ReadObject(context.Request.InputStream) as MonitoringDataPoint[];
-                    var transmissionTime = long.Parse(context.Request.Headers["x-ms-qps-transmission-time"], CultureInfo.InvariantCulture);
+                    var transmissionTime = long.Parse(context.Request.Headers[QuickPulseConstants.XMsQpsTransmissionTimeHeaderName], CultureInfo.InvariantCulture);
 
                     this.samples.AddRange(dataPoints.Select(dp => Tuple.Create(new DateTimeOffset(transmissionTime, TimeSpan.Zero), dp)));
                 }