+++ Src/Web/Web.Shared.Net/Implementation/WebEventSource.cs
@@ -331,6 +331,30 @@
                 this.ApplicationName);
         }
 
+        [Event(
+            35,
+            Message = "Failed to hook onto AddOnSendingHeaders event. Exception {0}",
+            Level = EventLevel.Warning)]
+        public void HookAddOnSendingHeadersFailedWarning(string exception, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                35,
+                exception,
+                this.ApplicationName);
+        }
+
+        [Event(
+            36,
+            Message = "Failed to add target instrumentation key hash as a response header. Exception {0}",
+            Level = EventLevel.Warning)]
+        public void AddTargetHeaderFailedWarning(string exception, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                36,
+                exception,
+                this.ApplicationName);
+        }
+
         [NonEvent]
         private string GetApplicationName()
         {